+++ tests/conftest.py
@@ -1,5 +1,6 @@
 
 import os
+import shutil
 
 import pytest
 
@@ -10,10 +11,8 @@ import kordesii.decoders
 def strings_exe(tmpdir):
     """Creates and returns a copy of the strings.exe file in a temporary directory."""
     orig_path = os.path.join(os.path.dirname(kordesii.decoders.__file__), 'tests', 'strings.exe')
-    new_path = os.path.join(str(tmpdir), 'strings.exe')
-    with open(orig_path, 'rb') as orig_file:
-        with open(new_path, 'wb') as new_file:
-            new_file.write(orig_file.read())
+    new_path = tmpdir / 'strings.exe'
+    shutil.copy(orig_path, str(new_path))
     return new_path
 
 
@@ -21,8 +20,6 @@ def strings_exe(tmpdir):
 def Sample_decoder(tmpdir):
     """Creates and returns the Sample decoder stored in a temporary directory."""
     orig_path = os.path.join(os.path.dirname(kordesii.decoders.__file__), 'Sample.py')
-    new_path = os.path.join(str(tmpdir), 'Sample.py')
-    with open(orig_path, 'rb') as orig_file:
-        with open(new_path, 'wb') as new_file:
-            new_file.write(orig_file.read())
+    new_path = tmpdir / 'Sample.py'
+    shutil.copy(orig_path, str(new_path))
     return new_path